requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
selenium
webdriver-manager
google-generativeai
dotenv
numpy
scikit-learn
sentence-transformers

